3rd commit

+++ src/Calculator.java
@@ -1,7 +1,7 @@
 
 public class Calculator {
 	
-	public int add (int x , int y) {
+	public double add (double x , double y) {
 		return x+y ;
 	}
